+++ src/enums/VariableDataType.ts
@@ -1,0 +1,11 @@
+enum VariableDataType {
+  String = 'string',
+  Number = 'number',
+  Boolean = 'boolean',
+  Date = 'date',
+  Address = 'address',
+  Table = 'table',
+  List = 'list',
+}
+
+export default VariableDataType;